refactor: move opcode
