+++ tiny/src/compiler.rs
@@ -1,5 +1,5 @@
 use crate::ast::{BinaryOperation, Expr, Stmt};
-use crate::opcode::OpCode;
+use crate::value_object::opcode::OpCode;
 use crate::tiny_object::TinyObject;
 
 #[derive(Debug)]
