+++ tiny/src/value_object.rs
@@ -1,0 +1,1 @@
+pub mod opcode;
